+++ src/FluxFlow.Components.Designer/IComponentDesignMetadataProvider.cs
@@ -1,0 +1,8 @@
+using FluxFlow.Components.Designer.Contracts;
+
+namespace FluxFlow.Components.Designer;
+
+public interface IComponentDesignMetadataProvider
+{
+    IReadOnlyCollection<ComponentDesignMetadata> GetMetadata();
+}